Rename `serve.rs` to `pages.rs` for clarity.
